Revert "fix(network) extract NetworkSelector component"
This reverts commit a46f3f2256222b6e9e5c6963b148b46e738328d0.
